requirements.txt
alembic==1.4.3
amqp==2.6.1
aniso8601==8.0.0
argon2-cffi==20.1.0
billiard==3.6.3.0
blinker==1.4
boto3==1.16.25
botocore==1.19.25
celery==4.4.7
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==4.4.2
Flask==1.1.2
Flask-Argon2==0.2.0.0
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
flask-rest-paginate==1.1.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SendGrid==0.7.1
Flask-SQLAlchemy==2.4.4
future==0.18.2
humanize==3.2.0
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
kombu==4.6.11
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
numpy==1.19.3
pandas==1.1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.14
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-http-client==3.3.1
pytz==2020.4
PyVimeo==1.1.0
rabbitmq==0.2.0
requests==2.25.1
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.4
sendgrid==6.4.7
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Searchable==1.2.0
SQLAlchemy-Utils==0.36.8
starkbank-ecdsa==1.1.0
threadpoolctl==2.1.0
tinydb==4.3.0
tornado==6.1
tuspy==0.2.4
typing-extensions==3.7.4.3
urllib3==1.26.2
validators==0.18.2
vine==1.3.0
wcwidth==0.2.5
Werkzeug==1.0.1
Whoosh==2.7.4
zipp==3.4.0
